acq stitching: fix cyclic dependency
In [1]: import odemis.acq.stitching
---------------------------------------------------------------------------
ImportError                               Traceback (most recent call last)
<ipython-input-1-3e50e201fda9> in <module>()
----> 1 import odemis.acq.stitching

/home/piel/development/odemis/src/odemis/acq/stitching/__init__.py in <module>()
     19 from odemis.acq.stitching._constants import REGISTER_GLOBAL_SHIFT, REGISTER_SHIFT, \
     20     REGISTER_IDENTITY, WEAVER_MEAN, WEAVER_COLLAGE, WEAVER_COLLAGE_REVERSE
---> 21 from odemis.acq.stitching._tiledacq import acquireTiledArea, estimateTiledAcquisitionTime, estimateTiledAcquisitionMemory
     22 from odemis.acq.stitching._registrar import *
     23 from odemis.acq.stitching._weaver import *

/home/piel/development/odemis/src/odemis/acq/stitching/_tiledacq.py in <module>()
     26 from odemis import model, dataio
     27 from odemis.acq import acqmng
---> 28 from odemis.acq import stitching
     29 from odemis.acq.align.autofocus import MeasureOpticalFocus, AutoFocus, MTD_EXHAUSTIVE
     30 from odemis.acq.stitching import register, weave

ImportError: cannot import name stitching

On Python 2, it's not possible to use "from ... import ..." in a cycle,
even if it's just a module.
In any case, cyclic dependencies are muddy, so better break it.

=> Move the simple functions to a separate private module.
