+++ src/odemis/acq/stitching/_constants.py
@@ -24,4 +24,3 @@ REGISTER_GLOBAL_SHIFT = 2
 WEAVER_MEAN = 0
 WEAVER_COLLAGE = 1
 WEAVER_COLLAGE_REVERSE = 2
-
